Refactor integration tests
- Split them out into separate files
- Import them into the crate so private APIs can be used

Inspired by:
https://matklad.github.io/2021/02/27/delete-cargo-integration-tests.html
http://xion.io/post/code/rust-unit-test-placement.html
https://doc.rust-lang.org/reference/items/modules.html#the-path-attribute
